+++ src/main/kotlin/icu/windea/pls/ep/codeInsight/documentation/ParadoxQuickDocTextProvider.kt
@@ -6,22 +6,18 @@ import icu.windea.pls.lang.annotations.WithGameTypeEP
 import icu.windea.pls.lang.codeInsight.documentation.ParadoxDocumentationManager
 import icu.windea.pls.lang.codeInsight.documentation.ParadoxDocumentationTarget
 import icu.windea.pls.lang.psi.ParadoxPsiMatcher
-import icu.windea.pls.lang.psi.light.ParadoxComplexEnumValueLightElement
-import icu.windea.pls.lang.psi.light.ParadoxDynamicValueLightElement
-import icu.windea.pls.lang.psi.light.ParadoxParameterLightElement
-import icu.windea.pls.script.psi.ParadoxDefinitionElement
-import icu.windea.pls.script.psi.ParadoxScriptScriptedVariable
 
 /**
  * 用于为各种目标提供提示文本，并渲染到快速文档中。
  *
  * 目前仅支持以下类型的目标：
  *
- * - 封装变量（[ParadoxScriptScriptedVariable]）
- * - 定义（[ParadoxDefinitionElement]）
- * - 复杂枚举（[ParadoxComplexEnumValueLightElement]）
- * - 动态值（[ParadoxDynamicValueLightElement]）
- * - 参数（[ParadoxParameterLightElement]）
+ * - 封装变量（参见 [ParadoxQuickDocTextProviderBase.ScriptedVariable]）
+ * - 定义（参见 [ParadoxQuickDocTextProviderBase.Definition]）
+ * - 内联脚本（参见 [ParadoxQuickDocTextProviderBase.InlineScript]）
+ * - 复杂枚举（参见 [ParadoxQuickDocTextProviderBase.ComplexEnumValue]）
+ * - 动态值（参见 [ParadoxQuickDocTextProviderBase.DynamicValue]）
+ * - 参数（参见 [ParadoxQuickDocTextProviderBase.Parameter]）
  *
  * @see ParadoxDocumentationTarget
  * @see ParadoxDocumentationManager