+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/hints/script/ParadoxComplexEnumValueHintTextHintsProvider.kt
@@ -21,7 +21,7 @@ import icu.windea.pls.script.psi.ParadoxScriptStringExpressionElement
 
 /**
  * 通过内嵌提示显示复杂枚举值的提示文本。
- * 来自本地化名称（即同名的本地化），或者对应的扩展规则。优先级从低到高。
+ * 来自显示名称（即同名的本地化文本），或者对应的扩展规则。优先级从低到高。
  *
  * @see ParadoxHintTextProvider
  * @see ParadoxHintTextProviderBase.ComplexEnumValue