+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/hints/script/ParadoxDynamicValueHintTextHintsProvider.kt
@@ -22,7 +22,7 @@ import icu.windea.pls.script.psi.isExpression
 
 /**
  * 通过内嵌提示显示动态值的提示文本。
- * 来自本地化名称（即同名的本地化），或者对应的扩展规则。优先级从低到高。
+ * 来自显示名称（即同名的本地化文本），或者对应的扩展规则。优先级从低到高。
  *
  * @see ParadoxHintTextProvider
  * @see ParadoxHintTextProviderBase.DynamicValue